Move the Course interface to its own file
